- notable games section complete

+++ controllers/MembersController.php
@@ -82,9 +82,11 @@
                         $game = new NotableGame();
                         $game->player_id = $member->id;
                         $game->lichess_id = substr($matches[1], 0, 8);
-                        if( !$game->save() ){
+                        $game->description = strip_tags(Yii::$app->request->post('gamedesc'));
+                        if( !$game->save() ) {
                             throw new HttpException(400, 'Failed to save game');
                         }
+                        return $game->id;
                     }
                     else {
                         throw new HttpException(400, 'Invalid lichess game url');
@@ -95,6 +97,9 @@
             }
 
             if( Yii::$app->request->isDelete ) {
+                if( Yii::$app->user->isGuest || Yii::$app->user->identity->getId() != $member->id ) {
+                    throw new HttpException(403, "Can't delete this");
+                }
                 $id = Yii::$app->request->bodyParams['gid'];
                 $game = NotableGame::findOne($id);
                 $game->delete();
